requirements.txt
optimum[onnxruntime]
nicegui
fastapi[standard]
slowapi
authlib
starlette
slack-bolt
pydantic
retry
pyjwt
inflect
humanize
memoization
azure-monitor-opentelemetry
tomli
icecream
python-dotenv